feat: modularize repository into a service-oriented architecture
- Restructured the codebase into core, services, and pipelines.
- Isolated VLA, Image Tagging, Segmentation, and Synthetic Data into independent services.
- Added pyproject.toml for each service to manage dependencies independently.
- Updated README to reflect the new architecture and CI/CD strategy.
- Standardized package structure with __init__.py files.
