+++ services/image-tagging/pyproject.toml
@@ -1,0 +1,15 @@
+[project]
+name = "image-tagging"
+version = "0.1.0"
+description = "VLM-based image tagging service"
+dependencies = [
+    "torch",
+    "transformers",
+    "requests",
+    "pillow",
+    "qwen-vl-utils", # Often used with Qwen2-VL
+]
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
