+++ services/image-tagging/src/VLM.py
