+++ services/image-tagging/src/frame_annotator.py
