+++ services/image-tagging/src/hf_model.py
