+++ services/image-tagging/src/parsing.py
