+++ services/image-tagging/src/post_annotation.py
