+++ services/image-tagging/src/prompts.py
