+++ services/image-tagging/src/qwen_subtask_annotator.py
