+++ services/image-tagging/src/qwen_vlm_image.py
