+++ services/image-tagging/src/vlm_qwen.py
