+++ services/segmentation/pyproject.toml
@@ -1,0 +1,14 @@
+[project]
+name = "segmentation"
+version = "0.1.0"
+description = "Segmentation service using SAM3"
+dependencies = [
+    "torch",
+    "numpy",
+    "opencv-python",
+    "pillow",
+]
+
+[build-system]
+requires = ["setuptools>=61.0"]
+build-backend = "setuptools.build_meta"
