+++ services/synthetic-data/src/corner_case_tool.py
