+++ services/synthetic-data/src/image_prompt_tool.py
