+++ services/vla-inference/check_memory.py
